refactor(web): inline the token usage table cell
Render the badge and optional breakdown tooltip inside createTokenUsageTableColumn, matching the other column creators.
