+++ web/src/components/design-system/table/columns/createTokenUsageTableColumn.tsx
@@ -1,10 +1,9 @@
 import { type CellContext, type RowData } from "@tanstack/react-table";
+import { InfoIcon } from "lucide-react";
 
-import {
-  TokenUsageTableCell,
-  type TokenUsageTableCellProps,
-} from "@/src/components/design-system/table/components/TokenUsageTableCell/TokenUsageTableCell";
+import { TokenUsageBadge } from "@/src/components/token-usage-badge";
 import { Skeleton } from "@/src/components/ui/skeleton";
+import { BreakdownTooltip } from "@/src/features/traces/components/BreakdownTooltip";
 import {
   createTableColumn,
   type TableColumnOptions,
@@ -12,9 +11,27 @@ import {
 
 const tokenUsageLoadingCell = <Skeleton className="h-4 w-1/2" />;
 
+type TokenUsageDetails = Record<string, number | undefined>;
+
+type TokenUsageCounts = {
+  inputUsage: number;
+  outputUsage: number;
+  totalUsage: number;
+};
+
 export type TokenUsageTableColumnCell =
   | { type: "loading" }
-  | ({ type: "usage" } & TokenUsageTableCellProps);
+  | ({ type: "usage" } & TokenUsageCounts &
+      (
+        | {
+            details: TokenUsageDetails | TokenUsageDetails[];
+            pricingTierName?: string;
+          }
+        | {
+            details?: undefined;
+            pricingTierName?: undefined;
+          }
+      ));
 
 export function createTokenUsageTableColumn<
   TData extends RowData,
@@ -36,8 +53,28 @@ export function createTokenUsageTableColumn<
       if (!cell) return null;
       if (cell.type === "loading") return tokenUsageLoadingCell;
 
-      const { type: _type, ...props } = cell;
-      return <TokenUsageTableCell {...props} />;
+      const badge = (
+        <TokenUsageBadge
+          inputUsage={cell.inputUsage}
+          outputUsage={cell.outputUsage}
+          totalUsage={cell.totalUsage}
+          inline
+        />
+      );
+
+      if (!cell.details) return badge;
+
+      return (
+        <BreakdownTooltip
+          details={cell.details}
+          pricingTierName={cell.pricingTierName}
+        >
+          <div className="flex items-center gap-1">
+            {badge}
+            <InfoIcon className="h-3 w-3" />
+          </div>
+        </BreakdownTooltip>
+      );
     },
   });
 }
